Update test definitions
- Use testcase verify qualifications, instead of bare assert statements. That
way, all problematic values are shown at once, and the test results show clearly
which tests had an unexpected error, and which one had an incorrect value.
- Use a default relative tolerance of 1e-10, instead of an absolute tolerance
of 1e-4. This should be a tolerance that can be achieved in any double precision
calculation, except for special cases. Those should then not use the default.
- Rename the test object that is the first input of the Test functions to
testCase, to be in line with the mathworks documentation examples (and de facto
language standard).